ADV managers most recent summary data

This function returns abbreviated ADV data for all filing managers for the most recent period.

adv_managers_current_period_summary(select_names = c("dateDataADV",
  "isExempt", "idRegionSEC", "idCRD", "idSEC", "typeRegulationSEC",
  "nameEntityManager", "nameEntityManagerLegal", "addressOfficePrimary",
  "addressStreet1OfficePrimary", "addressStreet2OfficePrimary",
  "cityOfficePrimary", "stateOfficePrimary", "countryOfficePrimary",
  "zipOfficePrimary", "phoneOfficePrimary", "statusSEC", "dateStatusSEC",
  "dateADVLatest", "urlManager", "isForeignRegisteredEntity",
  "stateDateJurisdictionNotice", "idCIK", "hasAUMGreater1B", "idLEI",
  "hasAUMGreater100M",      "typeEntity", "countryEntityOrganized",
  "countEmployeesTotal", "countEmployeesInvestmentAdvisory",
  "amountAUMTotal", "amountAUMDiscretionary", "amountAUMNonDiscretionary",
  "countAccountsDiscretionary", "countAccountsNonDiscretionary",
  "countAccountsTotal", "isManagerSecuritiesPortfolio", "hasFeeAUM",
  "hasFeeHourlyCharge", "hasFeeSubscription", "hasFeeFixed",
  "hasFeeCommission", "hasFeePerformance", "hasFeeOther", "typeFeeOther",
  "isBrokerDealer", "isBrokerDealerRepresentative",
  "isCommodityPoolOperator",      "isFuturesMerchant",
  "isRealEstateBrokerDealerAgent", "isInsuranceBrokerAgent", "isBank",
  "isTrustCompany", "isRegisteredMunicipalAdviser",
  "isRegisteredSecuritySwapDealer", "isRegistredSecuritySwapParticipant",
  "isAccountingFirm", "isLawFirm", "isOtherFinancialProductSalesperson",
  "typeOtherFinancialProductSalesperson", "countEmployeesBrokerDealer",
  "countEmployeesStateRegisteredInvestmentAdviser",
  "countEmployeesStateRegisteredInvestmentAdviserMultipleEntities",
  "countEmployeesLicensedInsuranceAgents",
  "countEmployeesSolicitAdvisoryClients", "hasFelonyPleaConviction",
  "hasFelonyCharge", "hasMisdemeanorPleaConviction"),
  return_message = TRUE)

Arguments

return_message
return a message after parsing data

file_directory
directory you want to save your data into, if none specified a temporary file will be created

folder_name
older you want the data to be downloaded into

remove_files
remove the files from the folders

empty_trash
empty the trash after being read into R

Value

a data frame

Details

For multiple periods and all information see adv_managers_periods_summaries

See also

Examples

# NOT RUN {
adv_managers_current_period_summary(select_names = c("dateDataADV", "isExempt", "idRegionSEC", "idCRD", "idSEC", "typeRegulationSEC", "nameEntityManager", "nameEntityManagerLegal", 'addressOfficePrimary', "addressStreet1OfficePrimary", "addressStreet2OfficePrimary", "cityOfficePrimary", "stateOfficePrimary", "countryOfficePrimary", "zipOfficePrimary", "phoneOfficePrimary", "statusSEC", "dateStatusSEC", "dateADVLatest", "urlManager", "isForeignRegisteredEntity", "stateDateJurisdictionNotice", "idCIK", "hasAUMGreater1B", "idLEI", "hasAUMGreater100M", "typeEntity", "countryEntityOrganized", "countEmployeesTotal", "countEmployeesInvestmentAdvisory", "amountAUMTotal", "amountAUMDiscretionary", "amountAUMNonDiscretionary", "countAccountsDiscretionary", "countAccountsNonDiscretionary", "countAccountsTotal", "isManagerSecuritiesPortfolio", "hasFeeAUM", "hasFeeHourlyCharge", "hasFeeSubscription", "hasFeeFixed", "hasFeeCommission", "hasFeePerformance", "hasFeeOther", "typeFeeOther", "isBrokerDealer", "isBrokerDealerRepresentative", "isCommodityPoolOperator", "isFuturesMerchant", "isRealEstateBrokerDealerAgent", "isInsuranceBrokerAgent", "isBank", "isTrustCompany", "isRegisteredMunicipalAdviser", "isRegisteredSecuritySwapDealer", "isRegistredSecuritySwapParticipant", "isAccountingFirm", "isLawFirm", "isOtherFinancialProductSalesperson", "typeOtherFinancialProductSalesperson", "countEmployeesBrokerDealer", "countEmployeesStateRegisteredInvestmentAdviser", "countEmployeesStateRegisteredInvestmentAdviserMultipleEntities", "countEmployeesLicensedInsuranceAgents", "countEmployeesSolicitAdvisoryClients", "hasFelonyPleaConviction", "hasFelonyCharge", "hasMisdemeanorPleaConviction"))
# }