Partnership waterfall and cash flows

This function is a variant of the calculate_cash_flow_waterfall function that also includes the calculation of the share of cash flow available to each member based upon the specified percentage equity splits and promote allocation

calculate_cash_flow_waterfall_partnership(dates = c("2016-09-01",
  "2017-08-31"), cash_flows = c(-1500000, 1.05e+08),
  working_capital = 2e+05, promote_structure = c("20 over 12",
  "30 over 20", "50 over 3.5x", "100 over 10x"),
  assign_to_environment = TRUE, general_partner_pct = 0.05,
  gp_promote_share = 1, unnest_data = FALSE,
  exclude_partnership_total = FALSE, distribution_frequency = NA,
  is_actual_360 = TRUE, widen_promote_structure = FALSE,
  bind_to_cf = FALSE, remove_zero_cols = TRUE,
  widen_waterfall = FALSE)

Arguments

dates
vector of dates in year-month-day format

cash_flows
vector of cash flows

working_capital
amount of working capital

promote_structure
vector of the promote structure

general_partner_pct
ercentage of capital provided by general partner

gp_promote_share
share of promote to general partner

unnest_data
unnest final results

exclude_partnership_total
exclude total columns

distribution_frequency
frequency of distribution

is_actual_360
TRUE is the rate of return actual 360

widen_promote_structure
TRUE widen the promote structure

bind_to_cf
TRUE bind results to data frame

remove_zero_cols
TRUE remove zero-value columns

widen_waterfall
TRUE returns waterfall in wide form

Value

data_frameO

See also

Examples

# NOT RUN {
calculate_cash_flow_waterfall_partnership(dates =c("2016-09-01", "2017-08-31"), cash_flows = c(-1500000, 105000000),
working_capital = 200000, promote_structure = c("20 over 12", "30 over 20", "50 over 3.5x", "100 over 10x"),
general_partner_pct = .05, gp_promote_share = 1, unnest_data = F,
exclude_partnership_total = F,
distribution_frequency = 'annually', is_actual_360 = TRUE,
widen_promote_structure = FALSE, bind_to_cf = FALSE, remove_zero_cols = TRU,
widen_waterfall = FALSE)
# }