Leveraged return metrics

This function returns data with the outputs from a "back-of-the-envelope" leveraged math calculation that can be used for a quick and dirty underwriting of a leveragd cash flow stream.

calculate_leverage_metrics(purchase_price = 0,
  capitalized_acquisition_costs = 0, capital_investment = 0, revenue = 0,
  expenses = 0, loan_to_cost = 0, interest_rate = 0,
  borrow_capital_investment = F, borrow_capitalized_costs = T,
  leverage_threshold = 0.95, is_interest_only = FALSE,
  interest_only_periods = 0, is_actual_360 = TRUE,
  amortization_years = 30, amortization_months = 0, term_years = 30,
  term_months = 0, pct_loan_fee = 0, return_wide = T,
  return_message = T)

Arguments

purchase_price
vector of purchase prices

capitalized_acquisition_costs
vector of capitalized acquisition costs

capital_investment
vector of capital investment

revenue
vector of revenue amounts

expenses
vector of expenses

loan_to_cost
vector of loan to cost

interest_rate
vector nterest rates

borrow_capital_investment
TRUE borrow capital investment

borrow_capitalized_costs
TRUE borrow capitalized costs

leverage_threshold
maximum leverage allowed

is_interest_only
TRUE has interest only periods

interest_only_periods
count of interest only periods

is_actual_360
TRUE is interest calculated on actual 360 basis

amortization_years
amortization in years

amortization_months
amortization additional months

term_years
term of the loan in years

term_months
term of the loan additional months

pct_loan_fee
loan fee in percent

return_wide
TRUE return data in wide form

return_message
TRUE return a message after data import

is_actual_360
TRUE interest calculated on actual/360 basis

Value

a data_frame

See also

Other calculation: calculate_cash_flow_dates, calculate_cash_flow_waterfall_partnership, calculate_cash_flow_waterfall, calculate_cash_flows_returns, calculate_days_accrued_pref, calculate_irr_periods, calculate_loan_payment, calculate_residual_valuation_cap_rates, calculate_residual_valuation_ebitda_multiples, calculate_share_proceeds, calculate_valuation_post_money, tidy_promote_structure

Other leveraged finance calculation: calculate_cash_flow_dates, calculate_cash_flow_waterfall_partnership, calculate_cash_flow_waterfall, calculate_cash_flows_returns, calculate_irr_periods, calculate_loan_payment

Examples

# NOT RUN {
calculate_leverage_metrics(purchase_price = c("13,000,000", "9,000,000"),
capitalized_acquisition_costs = "300,000",
capital_investment = "$100,000", revenue = "$1,200,000", expenses = "$400,000",
loan_to_cost = .75,interest_rate = .045,borrow_capital_investment = TRUE,
borrow_capitalized_costs = TRUE, leverage_threshold = .8, is_interest_only = FALSE,
interest_only_periods = 0, is_actual_360 = TRUE, amortization_years = 30,
amortization_months = 0, term_years = 10, term_months = 0, pct_loan_fee = 0, return_wide = TRUE,
return_message = TRUE)
# }