Residual value, EBITDA multiple method

This function calculates residual values based upon the specified inputs. See calculate_residual_valuation_cap_rates for capitalization rate method.

calculate_residual_valuation_ebitda_multiples(ebitda_multiples = c(5, 10,
  15, 20), ebitda = "$27,500,000", cost_of_sale = "5%",
  debt_balance = "$350,000,000", return_wide = T)

Arguments

ebitda_multiples
vector of EBITDA multiples in numeric or character

ebitda
vector of EBITDA in numeric or character numeric/currency form

cost_of_sale
vector of cost of sale in percent or character percent form

debt_balance
vector of anticipated debt balance at sale in numeric or character numeric/currency form

return_wide
return data in wide or long form

Value

a data_frame

See also

Examples

# NOT RUN {
calculate_residual_valuation_ebitda_multiples(ebitda_multiples = c(5, 10, 15, 20), ebitda = "$27,500,000", cost_of_sale = "5%", debt_balance = "$350,000,000", return_wide = T)
# }