Post money valuation splits

This function calculates the the post-money ownership splits of an entity for specified inputs.

calculate_valuation_post_money(pre_money_valuation = "$45,000,000",
  percent_sold = "10%", return_wide = T)

Arguments

pre_money_valuation
vector of of valuations in numeric or character

percent_sold
vector of of amount of business sold in percent or character percent form

return_wide
Return data in wide or long form

Value

a data_frame()

See also

Examples

# NOT RUN {
calculate_valuation_post_money(pre_money_valuation = "$45,000,000", percent_sold = "10%", return_wide = T)
# }