ADV managers most recent summary data

This function returns abbreviated ADV data for all filing managers for the most recent period.

get_data_adv_managers_current_period_summary(select_names = c("dateDataADV",
  "isExempt", "idRegionSEC", "idCRD", "idSEC", "typeRegulationSEC",
  "nameEntityManager", "nameEntityManagerLegal", "addressOfficePrimary",
  "addressStreet1OfficePrimary", "addressStreet2OfficePrimary",
  "cityOfficePrimary", "stateOfficePrimary", "countryOfficePrimary",
  "zipOfficePrimary", "phoneOfficePrimary", "statusSEC", "dateStatusSEC",
  "dateADVLatest", "urlManager", "isForeignRegisteredEntity",
  "stateDateJurisdictionNotice", "idCIK", "hasAUMGreater1B", "idLEI",
  "hasAUMGreater100M",      "typeEntity", "countryEntityOrganized",
  "countEmployeesTotal", "countEmployeesInvestmentAdvisory", "amountAUMTotal",
  "amountAUMDiscretionary", "amountAUMNonDiscretionary",
  "countAccountsDiscretionary", "countAccountsNonDiscretionary",
  "countAccountsTotal", "isManagerSecuritiesPortfolio", "hasFeeAUM",
  "hasFeeHourlyCharge", "hasFeeSubscription", "hasFeeFixed", "hasFeeCommission",
  "hasFeePerformance", "hasFeeOther", "typeFeeOther", "isBrokerDealer",
  "isBrokerDealerRepresentative", "isCommodityPoolOperator",
  "isFuturesMerchant", "isRealEstateBrokerDealerAgent",
  "isInsuranceBrokerAgent", "isBank", "isTrustCompany",
  "isRegisteredMunicipalAdviser", "isRegisteredSecuritySwapDealer",
  "isRegistredSecuritySwapParticipant", "isAccountingFirm", "isLawFirm",
  "isOtherFinancialProductSalesperson", "typeOtherFinancialProductSalesperson",
  "countEmployeesBrokerDealer",
  "countEmployeesStateRegisteredInvestmentAdviser",
  "countEmployeesStateRegisteredInvestmentAdviserMultipleEntities",
  "countEmployeesLicensedInsuranceAgents",
  "countEmployeesSolicitAdvisoryClients", "hasManagerFelonyPleaConviction",
  "hasManagerFelonyCharge", "hasManagerMisdemeanorPleaConviction"),
  return_message = TRUE)

Arguments

return_message
return a message after parsing data

file_directory
directory you want to save your data into, if none specified a temporary file will be created

folder_name
older you want the data to be downloaded into

remove_files
remove the files from the folders

empty_trash
empty the trash after being read into R

Value

a data frame

Details

For multiple periods and all information see get_data_adv_managers_periods_summariess

See also

Other IAPD: get_data_adv_managers_filings, get_data_adv_managers_metadata, get_data_adv_managers_periods_summaries, get_data_finra_entities, get_data_finra_people

Other ADV: get_data_adv_managers_filings, get_data_adv_managers_periods_summaries

Other entity search: get_data_adv_managers_filings, get_data_adv_managers_metadata, get_data_adv_managers_periods_summaries, get_data_finra_entities, get_data_nareit_entities, get_data_reit_funds, get_data_rf_leis, get_data_rf_sec_13F_companies, get_data_sec_bankruptcies, get_data_sec_broker_dealers, get_data_sec_ciks, get_data_sec_closed_end_funds, get_data_sec_cusips, get_data_sec_filer, get_data_sec_filing_entities, get_data_sec_investment_companies, get_data_sec_money_market_funds, get_data_sec_municipal_advisors, get_data_sec_xbrl_periods, get_data_securities_offerings, get_data_us_public_companies, get_data_ycombinator_alumni

Other fund data: get_data_adv_managers_filings, get_data_adv_managers_metadata, get_data_adv_managers_periods_summaries, get_data_reit_funds, get_data_sec_ciks, get_data_sec_closed_end_funds, get_data_sec_money_market_funds

Examples

# NOT RUN {
get_data_adv_managers_current_period_summary(select_names = c("dateDataADV", "isExempt", "idRegionSEC", "idCRD", "idSEC", "typeRegulationSEC", "nameEntityManager", "nameEntityManagerLegal", 'addressOfficePrimary', "addressStreet1OfficePrimary", "addressStreet2OfficePrimary", "cityOfficePrimary", "stateOfficePrimary", "countryOfficePrimary", "zipOfficePrimary", "phoneOfficePrimary", "statusSEC", "dateStatusSEC", "dateADVLatest", "urlManager", "isForeignRegisteredEntity", "stateDateJurisdictionNotice", "idCIK", "hasAUMGreater1B", "idLEI", "hasAUMGreater100M", "typeEntity", "countryEntityOrganized", "countEmployeesTotal", "countEmployeesInvestmentAdvisory", "amountAUMTotal", "amountAUMDiscretionary", "amountAUMNonDiscretionary", "countAccountsDiscretionary", "countAccountsNonDiscretionary", "countAccountsTotal", "isManagerSecuritiesPortfolio", "hasFeeAUM", "hasFeeHourlyCharge", "hasFeeSubscription", "hasFeeFixed", "hasFeeCommission", "hasFeePerformance", "hasFeeOther", "typeFeeOther", "isBrokerDealer", "isBrokerDealerRepresentative", "isCommodityPoolOperator", "isFuturesMerchant", "isRealEstateBrokerDealerAgent", "isInsuranceBrokerAgent", "isBank", "isTrustCompany", "isRegisteredMunicipalAdviser", "isRegisteredSecuritySwapDealer", "isRegistredSecuritySwapParticipant", "isAccountingFirm", "isLawFirm", "isOtherFinancialProductSalesperson", "typeOtherFinancialProductSalesperson", "countEmployeesBrokerDealer", "countEmployeesStateRegisteredInvestmentAdviser", "countEmployeesStateRegisteredInvestmentAdviserMultipleEntities", "countEmployeesLicensedInsuranceAgents", "countEmployeesSolicitAdvisoryClients", "hasManagerFelonyPleaConviction", "hasManagerFelonyCharge", "hasManagerMisdemeanorPleaConviction"))
# }