IAPD registered managers ADV data

This function parses the specified sections of the Form ADV for specified managers.

get_data_adv_managers_filings(entity_names = NULL, crd_ids = NULL,
  all_sections = TRUE, section_names = c("Registration",
  "Identifying Information", "Organization", "Successions",
  "Private Fund Reporting", "Direct Manager Owners", "Indirect Manager Owners",
  "Other Manager Information", "Manager Signatories"), flatten_tables = TRUE,
  gather_data = FALSE, assign_to_environment = TRUE)

Arguments

entity_names
vector entities you want to search

crd_ids
vector of CRDs to search

all_sections
include all sections TRUE(default) searches all sections

section_names
sections to search

Registration: registration information

Identifying Information: indentifying information

Organization: organizational structure information

Successions: succession information

Private Fund Reporting: private funds controlled by the manager

Direct Manager Owners: direct owners of the fund manager

Indirect Manager Owners: indirect owners of the fund manager

Other Manager Information: other information

Manager Signatories: manager signatories #'

flatten_tables
TRUE flattens data with multiple values into wide form

gather_data
TRUE returns a long data frame

assign_to_environment
TRUE assign individual data frames to your environment

Value

a data_frame

See also

Other IAPD: get_data_adv_managers_current_period_summary, get_data_adv_managers_metadata, get_data_adv_managers_periods_summaries, get_data_finra_entities, get_data_finra_people

Other ADV: get_data_adv_managers_current_period_summary, get_data_adv_managers_periods_summaries

Other entity search: get_data_adv_managers_current_period_summary, get_data_adv_managers_metadata, get_data_adv_managers_periods_summaries, get_data_finra_entities, get_data_nareit_entities, get_data_reit_funds, get_data_rf_leis, get_data_rf_sec_13F_companies, get_data_sec_bankruptcies, get_data_sec_broker_dealers, get_data_sec_ciks, get_data_sec_closed_end_funds, get_data_sec_cusips, get_data_sec_filer, get_data_sec_filing_entities, get_data_sec_investment_companies, get_data_sec_money_market_funds, get_data_sec_municipal_advisors, get_data_sec_xbrl_periods, get_data_securities_offerings, get_data_us_public_companies, get_data_ycombinator_alumni

Other fund data: get_data_adv_managers_current_period_summary, get_data_adv_managers_metadata, get_data_adv_managers_periods_summaries, get_data_reit_funds, get_data_sec_ciks, get_data_sec_closed_end_funds, get_data_sec_money_market_funds

Examples

# NOT RUN {
get_data_adv_managers_filings(entity_names = c('Blackstone Real Estate'), crd_ids = NULL,
 all_sections = TRUE,  section_names = NULL,
 flatten_tables = TRUE, gather_data = FALSE,
 assign_to_environment = TRUE)
# }