FINRA registered entities

This function returns information for any Financial Industry Regulatory Authority [FINRA] registered entity.

get_data_finra_entities(entity_names = NULL, ocr_pdf = TRUE,
  score_threshold = 0.2, return_message = TRUE)

Arguments

entity_names
vector of names to search

ocr_pdf
if TRUE returns OCR'd broker report PDF

score_threshold
matching score threshold for the search name if NULL there is no threshold

return_message
return a message upon parsing TRUE, FALS

Value

a data frame

Note

Use get_data_finra_people for registered people

References

FINRA

See also

Other FINRA: get_data_adv_managers_metadata, get_data_finra_people

Other IAPD: get_data_adv_managers_current_period_summary, get_data_adv_managers_filings, get_data_adv_managers_metadata, get_data_adv_managers_periods_summaries, get_data_finra_people

Other entity search: get_data_adv_managers_current_period_summary, get_data_adv_managers_filings, get_data_adv_managers_metadata, get_data_adv_managers_periods_summaries, get_data_nareit_entities, get_data_reit_funds, get_data_rf_leis, get_data_rf_sec_13F_companies, get_data_sec_bankruptcies, get_data_sec_broker_dealers, get_data_sec_ciks, get_data_sec_closed_end_funds, get_data_sec_cusips, get_data_sec_filer, get_data_sec_filing_entities, get_data_sec_investment_companies, get_data_sec_money_market_funds, get_data_sec_municipal_advisors, get_data_sec_xbrl_periods, get_data_securities_offerings, get_data_us_public_companies, get_data_ycombinator_alumni

Examples

# NOT RUN {
get_data_finra_entities(entity_names = c("EJF", "Blackstone", "Next Play", "Simple Capital"), score_threshold = .25,
ocr_pdf = TRUE)
# }