SEC filer

This is function imports data for a specified SEC filing entity. An SEC filing entity can be a person, public company or private filer. This function requires that the entity has a Central Index Key [CIK].

get_data_sec_filer(entity_names = NULL, tickers = NULL, ciks = NULL,
  tables = NULL, nest_data = FALSE, parse_all_filing_url_data = FALSE,
  parse_xbrl = FALSE, parse_subsidiaries = FALSE, parse_13F = FALSE,
  parse_asset_files = FALSE, parse_small_offerings = FALSE,
  parse_complete_text_filings = FALSE, parse_form_d = FALSE,
  parse_form_3_4s = FALSE, assign_to_environment = TRUE,
  return_message = TRUE)

Arguments

entity_names
vector names to search

tickers
character vector of ticker symbols to search

ciks
numeric vector of CIKs

tables
tables to include if they exist

General: general information about the filer

CIK Filings: summarised filings for a CIK

Filings: summarised filings for an entity, slightly different than CIK Filings

Private Offerings: parses any private offerings

Related Parties: parses any related parties [people]

Traders: parses major traders

C Level: parses information about executives

MDA: parses text from company 10-K Management Discussion and Analysis [MDA] section

Owners: parses information about major owners

Insider Trades: parses insider trade information

Trades: parses all trade information

nest_data
return a nested data frame TRUE, FALSE

parse_all_filing_url_data
TRUE parses every SEC fling link

parse_xbrl
TRUE parse XBRL for public companies, data starts in 2009

parse_subsidiaries
TRUE parse all filer subsidiaries (default)

parse_13F
TRUE parse 13F's for institutional managers

parse_asset_files
TRUE parses ABS XML for ABS Asset Data filing entities (default)

parse_small_offerings
TRUE parses Regulation CrowdFunding Form 1-A data if any exists for a filer (default)

assign_to_environment
true assigns individual data frames to your environment

return_message
TRUE return a message after data import

Value

where nest_data is TRUE a nested data_frame by asset, where nest_data is FALSE a data_frame

Details

The function acquires information for the specified tables and will auto parse forms if the filer has them and the user activates these parameters.

See also

Other SEC: get_data_edgar_filing_streams, get_data_mmf_owned_debt_securities, get_data_recent_insider_trades, get_data_rf_leis, get_data_rf_sec_13F_companies, get_data_sec_bankruptcies, get_data_sec_broker_dealers, get_data_sec_ciks, get_data_sec_closed_end_funds, get_data_sec_cusips, get_data_sec_failed_to_deliver_securities, get_data_sec_filing_entities, get_data_sec_foia_requests, get_data_sec_investment_companies, get_data_sec_money_market_funds, get_data_sec_municipal_advisors, get_data_sec_securities_filing_counts, get_data_sec_securities_metrics_by_exchange, get_data_securities_offerings, get_data_us_public_companies, get_dictionary_sec_filing_codes, get_dictionary_sec_form_codes, get_dictionary_sic_codes

Other Rank and Filed: get_data_mmf_owned_debt_securities, get_data_recent_insider_trades, get_data_rf_leis, get_data_rf_sec_13F_companies, get_data_sec_filing_entities, get_data_securities_offerings, get_data_us_public_companies

Other entity search: get_data_adv_managers_current_period_summary, get_data_adv_managers_filings, get_data_adv_managers_metadata, get_data_adv_managers_periods_summaries, get_data_finra_entities, get_data_nareit_entities, get_data_reit_funds, get_data_rf_leis, get_data_rf_sec_13F_companies, get_data_sec_bankruptcies, get_data_sec_broker_dealers, get_data_sec_ciks, get_data_sec_closed_end_funds, get_data_sec_cusips, get_data_sec_filing_entities, get_data_sec_investment_companies, get_data_sec_money_market_funds, get_data_sec_municipal_advisors, get_data_securities_offerings, get_data_us_public_companies, get_data_ycombinator_alumni

Other fund search: get_data_sec_investment_companies

Examples

# NOT RUN {
get_data_sec_filer(entity_names = 'HLT Holdco', tickers = c('FB'),
nest_data = TRUE, parse_subsidiaries = TRUE, parse_all_filing_url_data = TRUE,
parse_13F = TRUE, assign_to_environment = TRUE,
return_message = TRUE)

## Small Asset Filer Example

## ABS Example

#XBRL Example

# }