SEC CUSIPs

This function returns information on all Committee on Uniform Securities Identification Procedures [CUSIP] registered entities for a specified period. The SEC exposes this data in PDF form so this function can take a long time to run.

sec_cusips(only_most_recent = FALSE, years = 1996:2016,
  quarters = NULL, nest_data = FALSE, return_message = TRUE)

Arguments

only_most_recent
TRUE search only the most recent period

years
vector years to search, starting in 1996

quarters
quarters to search

NULL(default): all quarters,

1: Q1

2: Q2

3: Q3

4: Q4

nest_data
TRUE return nested data frame

return_message
TRUE return a message after data import

Value

nested data_frame or data_frame if nest_data = FALSE

References

The Securities and Exchange Commission

See also

Examples

# NOT RUN {
sec_cusips(only_most_recent = FALSE, years = c(2016, 2010),
quarters = c(4),
return_message = TRUE)
# }